% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_diagnostic_functions.R
\name{CorrelationPlot}
\alias{CorrelationPlot}
\title{Title}
\usage{
CorrelationPlot(mcmc_out, parnames, final_size = 2000, warmup_prop = 0)
}
\arguments{
\item{final_size}{}
}
\value{

}
\description{
Title
}
