% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_fit_functions.R
\name{FitOneVariableParamSpikeAndSlab_fixed_sigmagamma}
\alias{FitOneVariableParamSpikeAndSlab_fixed_sigmagamma}
\title{Title}
\usage{
FitOneVariableParamSpikeAndSlab_fixed_sigmagamma(data, nu_0 = 1e-04)
}
\arguments{
\item{nu_0}{}
}
\value{

}
\description{
Title
}
