% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_fit_functions.R
\name{FitOneVariableParamSpikeAndSlabDelay_marginal_delta}
\alias{FitOneVariableParamSpikeAndSlabDelay_marginal_delta}
\title{Title}
\usage{
FitOneVariableParamSpikeAndSlabDelay_marginal_delta(
  Cmcmc,
  nu_0 = 1e-04,
  niter = 5000
)
}
\arguments{
\item{niter}{}
}
\value{

}
\description{
Title
}
