% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_fit_functions.R
\name{FitOneVariableParamSpikeAndSlabDelay_marginal_delta_slow}
\alias{FitOneVariableParamSpikeAndSlabDelay_marginal_delta_slow}
\title{Title}
\usage{
FitOneVariableParamSpikeAndSlabDelay_marginal_delta_slow(
  data,
  nu_0 = 1e-04,
  niter = 5000
)
}
\arguments{
\item{nu_0}{}
}
\value{

}
\description{
Title
}
