% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_fit_functions.R
\name{OneVariableParamSpikeAndSlabDelay_marginal_delta_setup}
\alias{OneVariableParamSpikeAndSlabDelay_marginal_delta_setup}
\title{Title}
\usage{
OneVariableParamSpikeAndSlabDelay_marginal_delta_setup(
  data,
  nu_0 = 1e-04,
  niter = 5000
)
}
\arguments{
\item{nu_0}{}
}
\value{

}
\description{
Title
}
