% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_diagnostic_functions.R
\name{posterior_on_number_of_covariates_selected}
\alias{posterior_on_number_of_covariates_selected}
\title{Title}
\usage{
posterior_on_number_of_covariates_selected(mcmc.out, warmup_prop = 0.5)
}
\arguments{
\item{mcmc.out}{}
}
\value{

}
\description{
Title
}
