% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_diagnostic_functions.R
\name{posterior_predictive_estimate}
\alias{posterior_predictive_estimate}
\title{Title}
\usage{
posterior_predictive_estimate(
  data,
  mcmc_out,
  warmup_prop = 0.5,
  maxspecies = Inf
)
}
\arguments{
\item{l}{}
}
\value{

}
\description{
Title
}
