% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_diagnostic_functions.R
\name{PriorPosteriorPlotScalarParams}
\alias{PriorPosteriorPlotScalarParams}
\title{Title}
\usage{
PriorPosteriorPlotScalarParams(mcmc_out, parnames, warmup_prop = 0.5)
}
\arguments{
\item{warmup_prop}{}
}
\value{

}
\description{
Title
}
