% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{SimulateDataOneVariableParam}
\alias{SimulateDataOneVariableParam}
\title{Title}
\usage{
SimulateDataOneVariableParam(
  p = 5,
  N = 10,
  ni = 30,
  times = seq(0, 200, length.out = N),
  alpha = 15,
  beta = rbinom(n = p, size = 1, prob = 0.2),
  Gamma = 0.1,
  sigma = 0.1,
  param = "final size"
)
}
\arguments{
\item{sigma}{}
}
\value{

}
\description{
Title
}
