% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{SimulateDataOneVariableParam_reparam}
\alias{SimulateDataOneVariableParam_reparam}
\title{Title}
\usage{
SimulateDataOneVariableParam_reparam(
  J = 10,
  n = 30,
  times = seq(0, 200, length.out = J),
  beta = c(0.3, 0.3, 0, 0, 0, 0),
  Gamma = 0.1,
  sigma = 0.1,
  param = "delay",
  Lmax = 15,
  psi1 = 1,
  psi2 = 0.5,
  psi3 = 0.5
)
}
\arguments{
\item{psi3}{}
}
\value{

}
\description{
Title
}
